Keep shadow in its own file
